Switch to using tool-agnosic naming
Some references to Visual Studio Code remain in names of functions and
source files, but they should now be absent in the installed files and
environment variables.

The names of flag files could not be simply changed because they need
to be kept as-is on existing installations. Their name was made
configurable at build time, with the default set to the original
non-configurable name.

This fixes issue #3.